refactor: revert directory renaming change from "computing-unit-managing" back to "workflow-computing-unit"
