+++ frontend/src/app/common/service/computing-unit/computing-unit-actions/computing-unit-actions.service.ts
@@ -21,7 +21,7 @@ import { Injectable } from "@angular/core";
 import { Observable } from "rxjs";
 import { NzModalService } from "ng-zorro-antd/modal";
 import { ShareAccessComponent } from "../../../../dashboard/component/user/share-access/share-access.component";
-import { WorkflowComputingUnitManagingService } from "../computing-unit-managing/workflow-computing-unit-managing.service";
+import { WorkflowComputingUnitManagingService } from "../workflow-computing-unit/workflow-computing-unit-managing.service";
 import { DashboardWorkflowComputingUnit, WorkflowComputingUnitType } from "../../../type/workflow-computing-unit";
 import { NotificationService } from "../../notification/notification.service";
 import { unitTypeMessageTemplate } from "../../../util/computing-unit.util";
