+++ frontend/src/app/workspace/component/power-button/computing-unit-selection.component.ts
@@ -19,7 +19,7 @@
 
 import { ChangeDetectorRef, Component, OnInit } from "@angular/core";
 import { take } from "rxjs/operators";
-import { WorkflowComputingUnitManagingService } from "../../../common/service/computing-unit/computing-unit-managing/workflow-computing-unit-managing.service";
+import { WorkflowComputingUnitManagingService } from "../../../common/service/computing-unit/workflow-computing-unit/workflow-computing-unit-managing.service";
 import {
   DashboardWorkflowComputingUnit,
   WorkflowComputingUnitType,
